requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
MouseInfo==0.1.3
multidict==6.7.1
numpy==2.2.4
opencv-python==4.11.0.86
propcache==0.4.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySimpleGUI==5.0.8.2
pytweening==1.2.0
typing_extensions==4.15.0
yarl==1.22.0
